Install now supports using GIT or a provided library, added README, LICENSE, CHANGELOG and AUTHORS files. Modified the installation path to /usr/lib/prggmrlabs.
The provided library method is the prefered installation method as not
all users will have git installed on their systems.

The installation path is now /usr/lib/prggmrlabs this is being used
rather than /opt due to some issues with Mac OSX installation not
working properly with not being able to create the root level /opt
folder. The PHP /usr/lib/php/ path is not being because while prggmr is
written as a PHP library I consider it more of its own evironment within
PHP and the path can easily be changed during install if some feel it is
nessecary.